[REFACTOR] changed file struture
en: the files were divided between folders to help the
modularization and maintenance

pt: os arquivos foram divididos entre pastas para facilitar a
modularização e manutenção
